src/build_backend/post_requests/text_extraction/requirements.txt
requests
google-cloud-storage
PyPDF2
python-docx
Flask
Flask-Cors
gunicorn
